+++ packages/reason-relay/language-plugin/src/generator/ReservedKeywords.gen.tsx
@@ -1,0 +1,8 @@
+/* TypeScript file generated from ReservedKeywords.re by genType. */
+/* eslint-disable import/first */
+
+
+// tslint:disable-next-line:no-var-requires
+const ReservedKeywordsBS = require('./ReservedKeywords.bs');
+
+export const reservedKeywords: string[] = ReservedKeywordsBS.reservedKeywords;